refactor: separate the product-component
